server/requirements.txt
Flask
PyJWT
python-dotenv
Flask-SQLAlchemy